requirements.txt
# HDR Training Pipeline Requirements
# For RTX 3090 with CUDA 11.8+ support

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.8.0  # For alignment detection

# Progress bars
tqdm>=4.65.0

# Logging and visualization
tensorboard>=2.12.0

# Numerical computing
numpy>=1.24.0

